+++ sae-20190506/alibabacloud_sae20190506/models.py
@@ -12099,6 +12099,7 @@ class CreateApplicationRequest(TeaModel):
         service_tags: str = None,
         sidecar_containers_config: List[SidecarContainerConfig] = None,
         sls_configs: str = None,
+        startup_probe: str = None,
         termination_grace_period_seconds: int = None,
         timezone: str = None,
         tomcat_config: str = None,
@@ -12220,6 +12221,7 @@ class CreateApplicationRequest(TeaModel):
         self.sidecar_containers_config = sidecar_containers_config
         # [{"logDir":"","logType":"stdout"},{"logDir":"/tmp/a.log"}]
         self.sls_configs = sls_configs
+        self.startup_probe = startup_probe
         # 30
         self.termination_grace_period_seconds = termination_grace_period_seconds
         # Asia/Shanghai
@@ -12391,6 +12393,8 @@ class CreateApplicationRequest(TeaModel):
                 result['SidecarContainersConfig'].append(k.to_map() if k else None)
         if self.sls_configs is not None:
             result['SlsConfigs'] = self.sls_configs
+        if self.startup_probe is not None:
+            result['StartupProbe'] = self.startup_probe
         if self.termination_grace_period_seconds is not None:
             result['TerminationGracePeriodSeconds'] = self.termination_grace_period_seconds
         if self.timezone is not None:
@@ -12551,6 +12555,8 @@ class CreateApplicationRequest(TeaModel):
                 self.sidecar_containers_config.append(temp_model.from_map(k))
         if m.get('SlsConfigs') is not None:
             self.sls_configs = m.get('SlsConfigs')
+        if m.get('StartupProbe') is not None:
+            self.startup_probe = m.get('StartupProbe')
         if m.get('TerminationGracePeriodSeconds') is not None:
             self.termination_grace_period_seconds = m.get('TerminationGracePeriodSeconds')
         if m.get('Timezone') is not None:
@@ -12639,6 +12645,7 @@ class CreateApplicationShrinkRequest(TeaModel):
         service_tags: str = None,
         sidecar_containers_config_shrink: str = None,
         sls_configs: str = None,
+        startup_probe: str = None,
         termination_grace_period_seconds: int = None,
         timezone: str = None,
         tomcat_config: str = None,
@@ -12760,6 +12767,7 @@ class CreateApplicationShrinkRequest(TeaModel):
         self.sidecar_containers_config_shrink = sidecar_containers_config_shrink
         # [{"logDir":"","logType":"stdout"},{"logDir":"/tmp/a.log"}]
         self.sls_configs = sls_configs
+        self.startup_probe = startup_probe
         # 30
         self.termination_grace_period_seconds = termination_grace_period_seconds
         # Asia/Shanghai
@@ -12920,6 +12928,8 @@ class CreateApplicationShrinkRequest(TeaModel):
             result['SidecarContainersConfig'] = self.sidecar_containers_config_shrink
         if self.sls_configs is not None:
             result['SlsConfigs'] = self.sls_configs
+        if self.startup_probe is not None:
+            result['StartupProbe'] = self.startup_probe
         if self.termination_grace_period_seconds is not None:
             result['TerminationGracePeriodSeconds'] = self.termination_grace_period_seconds
         if self.timezone is not None:
@@ -13074,6 +13084,8 @@ class CreateApplicationShrinkRequest(TeaModel):
             self.sidecar_containers_config_shrink = m.get('SidecarContainersConfig')
         if m.get('SlsConfigs') is not None:
             self.sls_configs = m.get('SlsConfigs')
+        if m.get('StartupProbe') is not None:
+            self.startup_probe = m.get('StartupProbe')
         if m.get('TerminationGracePeriodSeconds') is not None:
             self.termination_grace_period_seconds = m.get('TerminationGracePeriodSeconds')
         if m.get('Timezone') is not None:
@@ -18741,6 +18753,7 @@ class DeployApplicationRequest(TeaModel):
         service_tags: str = None,
         sidecar_containers_config: List[SidecarContainerConfig] = None,
         sls_configs: str = None,
+        startup_probe: str = None,
         swimlane_pvtz_discovery_svc: str = None,
         termination_grace_period_seconds: int = None,
         timezone: str = None,
@@ -18999,6 +19012,7 @@ class DeployApplicationRequest(TeaModel):
         # 
         # > A Log Service project that is automatically created by SAE when you create an application is deleted when the application is deleted. Therefore, when you create an application, you cannot select a Log Service project that is automatically created by SAE for log collection.
         self.sls_configs = sls_configs
+        self.startup_probe = startup_probe
         self.swimlane_pvtz_discovery_svc = swimlane_pvtz_discovery_svc
         # The timeout period for a graceful shutdown. Default value: 30. Unit: seconds. Valid values: 1 to 300.
         self.termination_grace_period_seconds = termination_grace_period_seconds
@@ -19191,6 +19205,8 @@ class DeployApplicationRequest(TeaModel):
                 result['SidecarContainersConfig'].append(k.to_map() if k else None)
         if self.sls_configs is not None:
             result['SlsConfigs'] = self.sls_configs
+        if self.startup_probe is not None:
+            result['StartupProbe'] = self.startup_probe
         if self.swimlane_pvtz_discovery_svc is not None:
             result['SwimlanePvtzDiscoverySvc'] = self.swimlane_pvtz_discovery_svc
         if self.termination_grace_period_seconds is not None:
@@ -19347,6 +19363,8 @@ class DeployApplicationRequest(TeaModel):
                 self.sidecar_containers_config.append(temp_model.from_map(k))
         if m.get('SlsConfigs') is not None:
             self.sls_configs = m.get('SlsConfigs')
+        if m.get('StartupProbe') is not None:
+            self.startup_probe = m.get('StartupProbe')
         if m.get('SwimlanePvtzDiscoverySvc') is not None:
             self.swimlane_pvtz_discovery_svc = m.get('SwimlanePvtzDiscoverySvc')
         if m.get('TerminationGracePeriodSeconds') is not None:
@@ -19434,6 +19452,7 @@ class DeployApplicationShrinkRequest(TeaModel):
         service_tags: str = None,
         sidecar_containers_config_shrink: str = None,
         sls_configs: str = None,
+        startup_probe: str = None,
         swimlane_pvtz_discovery_svc: str = None,
         termination_grace_period_seconds: int = None,
         timezone: str = None,
@@ -19692,6 +19711,7 @@ class DeployApplicationShrinkRequest(TeaModel):
         # 
         # > A Log Service project that is automatically created by SAE when you create an application is deleted when the application is deleted. Therefore, when you create an application, you cannot select a Log Service project that is automatically created by SAE for log collection.
         self.sls_configs = sls_configs
+        self.startup_probe = startup_probe
         self.swimlane_pvtz_discovery_svc = swimlane_pvtz_discovery_svc
         # The timeout period for a graceful shutdown. Default value: 30. Unit: seconds. Valid values: 1 to 300.
         self.termination_grace_period_seconds = termination_grace_period_seconds
@@ -19873,6 +19893,8 @@ class DeployApplicationShrinkRequest(TeaModel):
             result['SidecarContainersConfig'] = self.sidecar_containers_config_shrink
         if self.sls_configs is not None:
             result['SlsConfigs'] = self.sls_configs
+        if self.startup_probe is not None:
+            result['StartupProbe'] = self.startup_probe
         if self.swimlane_pvtz_discovery_svc is not None:
             result['SwimlanePvtzDiscoverySvc'] = self.swimlane_pvtz_discovery_svc
         if self.termination_grace_period_seconds is not None:
@@ -20023,6 +20045,8 @@ class DeployApplicationShrinkRequest(TeaModel):
             self.sidecar_containers_config_shrink = m.get('SidecarContainersConfig')
         if m.get('SlsConfigs') is not None:
             self.sls_configs = m.get('SlsConfigs')
+        if m.get('StartupProbe') is not None:
+            self.startup_probe = m.get('StartupProbe')
         if m.get('SwimlanePvtzDiscoverySvc') is not None:
             self.swimlane_pvtz_discovery_svc = m.get('SwimlanePvtzDiscoverySvc')
         if m.get('TerminationGracePeriodSeconds') is not None:
@@ -21287,6 +21311,7 @@ class DescribeApplicationConfigResponseBodyData(TeaModel):
         service_tags: Dict[str, str] = None,
         sidecar_containers_config: List[DescribeApplicationConfigResponseBodyDataSidecarContainersConfig] = None,
         sls_configs: str = None,
+        startup_probe: str = None,
         swimlane_pvtz_discovery: str = None,
         tags: List[DescribeApplicationConfigResponseBodyDataTags] = None,
         termination_grace_period_seconds: int = None,
@@ -21570,6 +21595,7 @@ class DescribeApplicationConfigResponseBodyData(TeaModel):
         # 
         # If you do not need to modify the logging configurations when you deploy the application, configure the **SlsConfigs** parameter only in the first request. You do not need to include this parameter in subsequent requests. If you no longer need to use Log Service, leave the **SlsConfigs** parameter empty in the request.
         self.sls_configs = sls_configs
+        self.startup_probe = startup_probe
         self.swimlane_pvtz_discovery = swimlane_pvtz_discovery
         # The details of the tags.
         self.tags = tags
@@ -21811,6 +21837,8 @@ class DescribeApplicationConfigResponseBodyData(TeaModel):
                 result['SidecarContainersConfig'].append(k.to_map() if k else None)
         if self.sls_configs is not None:
             result['SlsConfigs'] = self.sls_configs
+        if self.startup_probe is not None:
+            result['StartupProbe'] = self.startup_probe
         if self.swimlane_pvtz_discovery is not None:
             result['SwimlanePvtzDiscovery'] = self.swimlane_pvtz_discovery
         result['Tags'] = []
@@ -22001,6 +22029,8 @@ class DescribeApplicationConfigResponseBodyData(TeaModel):
                 self.sidecar_containers_config.append(temp_model.from_map(k))
         if m.get('SlsConfigs') is not None:
             self.sls_configs = m.get('SlsConfigs')
+        if m.get('StartupProbe') is not None:
+            self.startup_probe = m.get('StartupProbe')
         if m.get('SwimlanePvtzDiscovery') is not None:
             self.swimlane_pvtz_discovery = m.get('SwimlanePvtzDiscovery')
         self.tags = []